Added nfa.clj, the main file for the backend.
This commit adds nfa.clj, the main file for the backend. This file contains all of the data structures that are used to store the NFA on the backend, as well as the functions for manipulating that state (adding/removing nodes/transitions, loading states, resetting the state, etc.) as well as functions for testing strings against the NFA (test-string and test-string-r).